fix: corrige a leitura e processamento de tweets
- Altera a função de semeadura para usar o gerador padrão do NumPy.
- Modifica a ordem de importação de colunas em constantes.
- Atualiza o carregamento de tweets para usar um DataFrame nomeado.
- Melhora a amostragem e o tratamento de erros ao ler CSVs.
- Refatora o splitter para calcular tamanhos de teste e validação corretamente.
- Adiciona tratamento de exceções mais informativo em clientes LLM e modelos de embeddings.
- Renomeia variáveis para maior clareza em funções de pipeline.
- Ajusta a limpeza de tweets para remover URLs antes de emoticons.
- Remove código não utilizado em testes de limpeza.

+++ src/data/splitter.py
@@ -61,8 +61,8 @@ def stratified_split(
 
     for _, group in shuffled.group_by(label_col, maintain_order=True):
         n = group.height
-        n_test = int(round(n * test_size))
-        n_val = int(round(n * val_size))
+        n_test = round(n * test_size)
+        n_val = round(n * val_size)
         test_parts.append(group.slice(0, n_test))
         val_parts.append(group.slice(n_test, n_val))
         train_parts.append(group.slice(n_test + n_val))
